move all rayyan class code to init file
